+++ rbx_types/src/lib.rs
@@ -2,11 +2,16 @@
 #[macro_use]
 mod serde_util;
 
+#[cfg(feature = "legacy-compat")]
+mod legacy_compat;
+
+mod attributes;
 mod axes;
 mod basic_types;
 mod binary_string;
 mod brick_color;
 mod content;
+mod error;
 mod faces;
 mod lister;
 mod physical_properties;
@@ -14,11 +19,13 @@ mod referent;
 mod shared_string;
 mod variant;
 
+pub use attributes::*;
 pub use axes::*;
 pub use basic_types::*;
 pub use binary_string::*;
 pub use brick_color::*;
 pub use content::*;
+pub use error::*;
 pub use faces::*;
 pub use physical_properties::*;
 pub use referent::*;